Manager Movies, template Manager Movies, model, Service about Movies

+++ src/main/java/com/webphim/webphim/Reponsitory/EpisodesRepository.java
@@ -3,5 +3,5 @@ package com.webphim.webphim.Reponsitory;
 import com.webphim.webphim.Model.Episodes;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EpisodesRepository extends JpaRepository<Episodes, Integer> {
+public interface EpisodesRepository extends JpaRepository<Episodes, Long> {
 }
